:exclamation:  Adding new map VMF. :exclamation: Do not use this one on any server just yet.
We had to add a new map VMF as the old one has become corrupt and causes leaks out of solid brushes - it seems to be a bug with either VBSP, or Hammer itself. We found that creating a new VMF and copying+pasting the old brushes and entities into a new VMF resolves the problem.

But due to this, all minigame and boss coordinates are now out of sync. Due to this, if you run the vmf on your own server, the gamemode will not properly work at all.