Initial sdk plugin
This plugin will handle certain dev tasks like calculating lengths of music used in the bumpers and other minigames / bosses.

The intention of the plugin is to only be run on a dev server, and executed via console commands.

The end goal of this plugin is to remove the soundlib dependency on the main plugin