Initial code for checking dxlevel
Not sure if this will actually work, needs testing. In theory, we might be able to check this bool on every player when we want to set an overlay, then not use overlay if bool is true. This means for System overlays (Speed changes, Win,lose, gameover, special round) we could just fallback to legacy hud text.