Remove mtf2_use_server_map_timelimit
Instead, just use mp_timelimit. If you set mp_timelimit to 0, then the gamemode will rely on mtf2_maxrounds to control when to end the gamemode. If you set mp_timelimit to a positive value (i.e. above 0, for example 10), then the gamemode will switch from a set number of rounds to play, to instead just play for the remaining timelimit on the map.