#232 Fix bossgame7 not setting antiflood time when being played
This will set the sm_flood_time convar to -1 when this bossgame runs, and revert it back to the original value once the bossgame finishes.

Tried to use the OnClientFloodCheck forward, but whilst that lets you return false, the highest return wins; i.e. antiflood.smx says the client is flooding but the gamemode says its not, antiflood wins. For now, this ConVar approach should work fine